Refs 8330. Move DDS examples to separate file
