+++ code/CMakeLists.txt
@@ -163,7 +163,7 @@ if(BUILDCOP)
     ####################################################################################################################
     # Compile example
     ####################################################################################################################
-    add_executable(doctest CodeTester.cpp)
+    add_executable(doctest CodeTester.cpp DDSCodeTester.cpp)
     target_include_directories(doctest PRIVATE ${PROJECT_BINARY_DIR}/external/eprosima/src/fastrtps/src/cpp
         ${PROJECT_BINARY_DIR}/external/eprosima/src/fastrtps/thirdparty/tinyxml2)
     target_link_libraries(doctest fastrtps)
