+++ docs/spelling_wordlist.txt
@@ -56,6 +56,7 @@ deserialize
 deserializes
 destructor
 diffie
+domainId
 doxygen
 eProsima
 endian
